+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # PyGPSClient Release Notes
 
+### RELEASE 1.4.25
+
+ENHANCEMENTS:
+
+1. Enhance scatterplot widget to extend zoom range to 0.01m and add optional fixed reference position. Delivers #160 and #161.
+1. Minor improvements to datalogging and track recording.
+
 ### RELEASE 1.4.24
 
 FIXES: